move(auth): `app/login/action.ts` → `app/actions/auth.ts`
